requirements.txt
opencv-python
kafka-python
apache-flink 
pillow
torch
torchvision
transformers
matplotlib
